requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
ca-certificates=2019.5.15=0
certifi=2019.6.16=py37_0
cycler=0.10.0=py37_0
descartes=1.1.0=py37_0
django=2.2.3=py37_0
django-filter=2.1.0=py_0
django-registration-redux=2.6=pypi_0
djangorestframework=3.9.4=py_0
freetype=2.9.1=hb4e5f40_0
geos=3.7.1=h0a44026_0
intel-openmp=2019.4=233
kiwisolver=1.1.0=py37h0a44026_0
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
libgfortran=3.0.1=h93005f0_2
libpng=1.6.37=ha441bb4_0
matplotlib=3.1.0=py37h54f8f79_0
mkl=2019.4=233
mkl_fft=1.0.12=py37h5e564d8_0
mkl_random=1.0.2=py37h27c97d8_0
ncurses=6.1=h0a44026_1
numpy=1.16.4=py37hacdab7b_0
numpy-base=1.16.4=py37h6575580_0
openssl=1.1.1c=h1de35cc_1
pandas=0.24.2=py37h0a44026_0
pip=19.1.1=py37_0
pyparsing=2.4.0=py_0
python=3.7.3=h359304d_0
python-dateutil=2.8.0=py37_0
pytz=2019.1=py_0
readline=7.0=h1de35cc_5
scipy=1.2.1=py37h1410ff5_0
setuptools=41.0.1=py37_0
shapely=1.6.4=py37he8793f5_0
six=1.12.0=py37_0
sqlite=3.28.0=ha441bb4_0
sqlparse=0.3.0=py_0
tk=8.6.8=ha441bb4_0
tornado=6.0.3=py37h1de35cc_0
wheel=0.33.4=py37_0
xz=5.2.4=h1de35cc_4
zlib=1.2.11=h1de35cc_3
